[ghstack-poisoned]

+++ aten/src/ATen/native/cuda/jit_utils.h
@@ -68,7 +68,7 @@ inline int can_vectorize_up_to(size_t default_alignment, void *pointer) {
 
 inline int can_vectorize_up_to(const KernelDescriptor &desc, c10::ArrayRef<char*> pointers) {
   TORCH_INTERNAL_ASSERT(desc.nOutputs == 1);
-  TORCH_INTERNAL_ASSERT(pointers.size() == 1 + desc.nInputs);
+  TORCH_INTERNAL_ASSERT(static_cast<int64_t>(pointers.size()) == 1 + desc.nInputs);
 
   // Deals with output
   auto result_size = c10::scalarTypeToTypeMeta(desc.result_type).itemsize();
